overfill protection added (for lems & coords)

+++ find_way.h
@@ -6,7 +6,7 @@
 /*   By: uhand <uhand@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/16 18:06:04 by uhand             #+#    #+#             */
-/*   Updated: 2019/09/17 12:01:08 by uhand            ###   ########.fr       */
+/*   Updated: 2019/09/18 12:38:15 by uhand            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,9 +27,9 @@ typedef struct			s_way
 	int					len_way;
 	int					num_way;
 	int					status;
-	int					last_steps;
-	int					steps;
-	int					full_steps;
+	long long			last_steps;
+	long long			steps;
+	long long			full_steps;
 	t_dllist			*rooms;
 	struct s_way		*next;
 	struct s_way		*prev;
@@ -41,8 +41,8 @@ typedef struct			s_way
 
 struct			s_way_set
 {
-	int					steps;
-	int					full_steps;
+	long long			steps;
+	long long			full_steps;
 	int					ways_cnt;
 	t_list_rooms		*rooms;
   	t_list_links		*links;
@@ -80,10 +80,10 @@ struct					s_find_way
 typedef struct			s_count_steps
 {
 	t_way				*ptr;
-	int					i;
-	int					pre_lems;
-	int					max_len;
-	float				steps;
+	long long			i;
+	long long			pre_lems;
+	long long			max_len;
+	double				steps;
 }						t_count_steps;
 
 /*
@@ -132,8 +132,8 @@ typedef struct			s_form_set
 	t_way				*ptr;
 	int					counter;
 	int					i;
-	int					steps_count;
-	int					min_steps_count;
+	long long			steps_count;
+	long long			min_steps_count;
 	int					i_min;
 }						t_form_set;
